pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "huxbot"
version = "0.1.0"
description = "AI agent framework powered by Google ADK and LiteLLM"
requires-python = ">=3.11"
dependencies = [
    "google-adk>=1.5.0",
    "litellm>=1.30.0",
    "typer[all]>=0.9.0",
    "pydantic>=2.0.0",
    "python-telegram-bot>=21.0",
    "discord.py>=2.3.0",
    "aiohttp>=3.9.0",
    "pyyaml>=6.0",
    "python-frontmatter>=1.1.0",
    "streamlit>=1.30.0",
]

[project.optional-dependencies]
hardware = ["pyserial-asyncio>=0.6"]
dev = ["pytest", "pytest-asyncio", "ruff"]

[project.scripts]
huxbot = "huxbot.cli.commands:app"
